requirements.txt
# Python 3.10
# Create & activate a 3.10 virtualenv, then:
#   pip install -r requirements.txt
#
# Supabase / SQLAlchemy quick install (equivalent):
#   pip install python-dotenv sqlalchemy psycopg2-binary
# (Use psycopg2-binary on Windows; plain psycopg2 needs build tools.)

# --- Core web framework ---
Flask>=2.3.0,<4.0
flask-cors>=4.0.0
Werkzeug>=2.3.0

# --- Database ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# --- HTTP / API client ---
requests>=2.28.0

# --- Data / Excel handling ---
pandas>=2.0.0
openpyxl>=3.1.0

# --- PDF generation ---
reportlab>=4.0.0

# --- Environment variable loader (.env support) ---
python-dotenv>=1.0.0


